AWS-001
Clean unnecessary hooks and split Portfolio.tsx into reusable components.
