+++ types/sections.ts
@@ -1,0 +1,1 @@
+export type SectionKey = "intro" | "about" | "projects" | "career" | "tech-stack" | "contact";
